+++ .changeset/five-apples-jam.md
@@ -2,4 +2,5 @@
 '@leafygreen-ui/expandable-card': patch
 ---
 
-Quick fix to vertically center title
+Quick fix to vertically center title and updates internal file structure
+.
